Removed support for cutil library and header file.
Cutil was never intented to be used outside of the SDK.  The removal of this
code is in support of this.  The CUDA_SDK_ROOT_DIR will continue to be
supported, in case users wish to use this to find files in the SDK.  There are
also two examples of how to use CUDA_SDK_ROOT_DIR to find header files and
libraries if users so wish.