libarchive: Use one architecture for try-compiles
We use CHECK_TYPE_SIZE in libarchive to check for the existence of some
types.  For universal binary builds on the Mac, the size check can fail
if it is inconsistent across architectures.  However, we do not actually
need the size so it is safe to do the checks for only one architecture.

See issue #9913.