Name Cygwin DLLs with SOVERSION, not VERSION
Cygwin versions .dll files by putting the version number in the file
name.  Our fix to issue #3571 taught CMake to do this, but it used the
VERSION target property.  It is better to use the SOVERSION property
since that is the interface (rather than implementation) version.

Change based on patch from issue #10122.