Do not warn for unknown CTest UPDATE_TYPE
In the CTest module we previously warned if the source directory did not
contain known version control directories.  The message was:

  "CTest cannot determine repository type. Please set UPDATE_TYPE
   to 'cvs' or 'svn'. CTest update will not work."

This was confusing when building sources from a tarball.  Furthermore,
we now support many more version control tools.  This feature is now
mature enough that the warning causes confusion more than it provides
real help.  We simply remove it.